Reads a Google Sheet and notifies people
This program reads information from Google Sheets that records if payments have been made from individuals, and sends an email via SMTP to notify them if they are behind. My goal to finish this is to expand it for the user to be able to choose other spreadsheet/data files, send texts via a SMS gateway as an alternative, and to be able to write information to files instead of only reading them.